+++ internal/api/api.go
@@ -1,4 +1,4 @@
-package check
+package api
 
 import (
 	"context"
@@ -6,8 +6,8 @@ import (
 	"strings"
 	"sync"
 
-	"go.szostok.io/codeowners-validator/internal/ptr"
-	"go.szostok.io/codeowners-validator/pkg/codeowners"
+	"go.szostok.io/codeowners/internal/ptr"
+	"go.szostok.io/codeowners/pkg/codeowners"
 )
 
 type (
@@ -111,4 +111,20 @@ func (s *SeverityType) Unmarshal(in string) error {
 	}
 
 	return nil
+}
+
+func (s *SeverityType) Set(in string) error {
+	switch strings.ToLower(in) {
+	case "error", "err":
+		*s = Error
+	case "warning", "warn":
+		*s = Warning
+	default:
+		return fmt.Errorf("not a valid severity type: %q", in)
+	}
+	return nil
+}
+
+func (s *SeverityType) Type() string {
+	return "SeverityType"
 }
